+++ src/components/favoriteStoresRanking/FavoriteStoresTitle.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../css/favoriteStoresRanking.css";
+import "../../css/favoriteStoresRanking/favoriteStoresRanking.css";
 
 export const FavoriteStoresTitle: React.FC = () => (
   <h1 className="favorite-stores-title">Favorite Stores Ranking</h1>
