+++ src/css/icons/priceLevelIcons.css
